+++ api/lib/application/preHandlers/authorization.js
@@ -1,5 +1,5 @@
 import { NotFoundError } from '../http-errors.js';
-import * as certificationCourseRepository from '../../infrastructure/repositories/certification-course-repository.js';
+import * as certificationCourseRepository from '../../../src/certification/shared/infrastructure/repositories/certification-course-repository.js';
 import * as sessionRepository from '../../../src/certification/session/infrastructure/repositories/session-repository.js';
 
 const verifySessionAuthorization = async (request, h, dependencies = { sessionRepository }) => {
