+++ api/lib/application/usecases/checkUserIsMemberOfCertificationCenterSession.js
@@ -1,4 +1,4 @@
-import * as certificationCourseRepository from '../../infrastructure/repositories/certification-course-repository.js';
+import * as certificationCourseRepository from '../../../src/certification/shared/infrastructure/repositories/certification-course-repository.js';
 import * as sessionRepository from '../../../src/certification/session/infrastructure/repositories/session-repository.js';
 
 const execute = async function ({
