+++ api/lib/application/usecases/checkUserOwnsCertificationCourse.js
@@ -1,4 +1,4 @@
-import * as certificationCourseRepository from '../../infrastructure/repositories/certification-course-repository.js';
+import * as certificationCourseRepository from '../../../src/certification/shared/infrastructure/repositories/certification-course-repository.js';
 
 const execute = async function ({ userId, certificationCourseId, dependencies = { certificationCourseRepository } }) {
   const certificationCourse = await dependencies.certificationCourseRepository.get(certificationCourseId);
