+++ api/lib/domain/services/verify-certificate-code-service.js
@@ -1,5 +1,5 @@
 import _ from 'lodash';
-import * as certificationCourseRepository from '../../../lib/infrastructure/repositories/certification-course-repository.js';
+import * as certificationCourseRepository from '../../../src/certification/shared/infrastructure/repositories/certification-course-repository.js';
 import { CertificateVerificationCodeGenerationTooManyTrials } from '../../../lib/domain/errors.js';
 import { config } from '../../config.js';
