+++ api/scripts/certification/launch-auto-jury-for-session.js
@@ -6,7 +6,7 @@ import { SessionFinalized } from '../../lib/domain/events/SessionFinalized.js';
 import * as certificationAssessmentRepository from '../../lib/infrastructure/repositories/certification-assessment-repository.js';
 import * as challengeRepository from '../../src/shared/infrastructure/repositories/challenge-repository.js';
 import * as certificationIssueReportRepository from '../../src/certification/shared/infrastructure/repositories/certification-issue-report-repository.js';
-import * as certificationCourseRepository from '../../lib/infrastructure/repositories/certification-course-repository.js';
+import * as certificationCourseRepository from '../../src/certification/shared/infrastructure/repositories/certification-course-repository.js';
 import { handleAutoJury } from '../../lib/domain/events/handle-auto-jury.js';
 import * as events from '../../lib/domain/events/index.js';
 import { logger } from '../../lib/infrastructure/logger.js';
